refactor: репозиторий разбит на источники данных
Мок реализация local-source данных вынесена в отдельный класс.
Репозиторий теперь только получает и передает данные
