+++ app/src/main/java/com/pasha/ytodo/presentation/MainActivity.kt
@@ -10,7 +10,6 @@ import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.pasha.ytodo.R
-import com.pasha.ytodo.data.repositories.TodoItemsRepositoryTestImpl
 import com.pasha.ytodo.domain.repositories.TodoItemRepositoryProvider
 import com.pasha.ytodo.domain.repositories.TodoItemsRepository
 import kotlinx.coroutines.flow.collect
